requirements.txt
PyJWT==2.10.1
Flask==3.1.2
redis=7.0.0
python-dotenv==1.2.1
